chore: abstract beautifystatnumber
